refactor: 육각형 아키텍쳐를 응집도 있게 수정
