--- user-domain/src/main/kotlin/com/xquare/v1userservice/user/api/CreateUserApi.kt
@@ -1,8 +1,0 @@
-package com.xquare.v1userservice.user.api
-
-import com.xquare.v1userservice.user.User
-import com.xquare.v1userservice.user.api.dtos.CreatUserDomainRequest
-
-interface CreateUserApi {
-    suspend fun saveUser(creatUserDomainRequest: CreatUserDomainRequest): User
-}
