--- user-domain/src/main/kotlin/com/xquare/v1userservice/user/api/GetUserDeviceTokensApi.kt
@@ -1,8 +1,0 @@
-package com.xquare.v1userservice.user.api
-
-import com.xquare.v1userservice.user.api.dtos.UserDeviceTokenResponse
-import java.util.UUID
-
-interface GetUserDeviceTokensApi {
-    suspend fun getUserDeviceTokensByIdIn(idList: List<UUID>): UserDeviceTokenResponse
-}
