--- user-domain/src/main/kotlin/com/xquare/v1userservice/user/api/GetUserInformationService.kt
@@ -1,9 +1,0 @@
-package com.xquare.v1userservice.user.api
-
-import com.xquare.v1userservice.user.User
-import java.util.UUID
-
-interface GetUserInformationService {
-    suspend fun getUserById(userId: UUID): User
-    suspend fun getUserByAccountId(accountId: String): User
-}
