--- user-domain/src/main/kotlin/com/xquare/v1userservice/user/api/impl/GetUserDeviceTokensApiImpl.kt
@@ -1,20 +1,0 @@
-package com.xquare.v1userservice.user.api.impl
-
-import com.xquare.v1userservice.annotations.DomainService
-import com.xquare.v1userservice.user.api.GetUserDeviceTokensApi
-import com.xquare.v1userservice.user.api.dtos.UserDeviceTokenResponse
-import com.xquare.v1userservice.user.spi.UserRepositorySpi
-import java.util.UUID
-
-@DomainService
-class GetUserDeviceTokensApiImpl(
-    private val userRepositorySpi: UserRepositorySpi
-) : GetUserDeviceTokensApi {
-    override suspend fun getUserDeviceTokensByIdIn(idList: List<UUID>): UserDeviceTokenResponse {
-        val userDeviceTokens = userRepositorySpi.findAllByIdIn(idList)
-            .filter { it.deviceToken != null }
-            .map { it.deviceToken!! }
-
-        return UserDeviceTokenResponse(userDeviceTokens)
-    }
-}
